refactor: simplify theme constants
