+++ src/client/ui/style/index.ts
@@ -19,7 +19,7 @@
  * @dependencies
  *   @rbxts/fusion ^0.4.0
  */
-export * from "./ThemeState";
+export * from "./theme";
 export * from "./gradients";
 export * from "./layouts";
 export * from "./sizes";
